requirements.txt
flask==3.1.3
flask-cors==6.0.2
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-api-python-client==2.193.0
